% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{check_job}
\alias{check_job}
\title{Check status of a compliance job}
\usage{
check_job(job_id, bearer_token = get_bearer(), verbose = TRUE)
}
\arguments{
\item{job_id}{A string containing the ID of the compliance job that will be checked.}

\item{bearer_token}{API bearer token.}

\item{verbose}{If FALSE, the function will not print any messages.}
}
\value{
If `verbose` is set to FALSE, return the jobs download URL if ready or FALSE if still in progress.
}
\description{
This function checks the status of a given compliance job.
}
