% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{combine_jobs}
\alias{combine_jobs}
\title{Start compliance job}
\usage{
combine_jobs(all_tweets, compliance_tweets, compliance_users)
}
\arguments{
\item{all_tweets}{Currently only "tweets" is allowed here.}

\item{compliance_tweets}{If FALSE the function will not print any messages.}

\item{compliance_users}{API bearer token.}
}
\description{
Start compliance job
}
