% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twCompliance.R
\name{download_compliance_check}
\alias{download_compliance_check}
\title{Download a compliacnce check}
\usage{
download_compliance_check(job_id, return_raw = FALSE)
}
\arguments{
\item{job_id}{The object generated by `start_compliance_check()`.}

\item{return_raw}{bool Should a combinded output be returned or the raw API response?}
}
\value{
data.frame The downloaded and combinded compliance check
}
\description{
This function downloads a compliance check
}
