% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twCompliance.R
\name{list_compliance_jobs}
\alias{list_compliance_jobs}
\title{List all compliance jobs}
\usage{
list_compliance_jobs(type = "all", status = "all", bearer_token = get_bearer())
}
\arguments{
\item{type}{Type of the job. Either 'all' (default), 'tweets', or 'users'.}

\item{status}{The status of the job. Either 'all' (default), 'in_progress', or 'complete'.}

\item{bearer_token}{API bearer token.}
}
\description{
This function prints a list of all compliance jobs.
}
