% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{list_jobs}
\alias{list_jobs}
\title{List all compliance jobs}
\usage{
list_jobs(type = "tweets", bearer_token = get_bearer())
}
\arguments{
\item{type}{Only tweets is currently supported.}

\item{bearer_token}{API bearer token.}
}
\description{
This function prints a list of all compliance jobs.
}
