% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{print.twCompliance}
\alias{print.twCompliance}
\title{Prints a twCompliance object}
\usage{
\method{print}{twCompliance}(x, ...)
}
\arguments{
\item{x}{A twCompliance object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints a twCompliance object
}
