% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twCompliance.R
\name{report_compliance}
\alias{report_compliance}
\title{Report deleted content}
\usage{
report_compliance(df, plot = TRUE)
}
\arguments{
\item{df}{A dataframe generated by `combine_status()`.}

\item{plot}{If `TRUE` plot a report as well.}
}
\value{
string ID of the compliance job. Save this to check the jobs status and download the results.
}
\description{
This function generates a report of deleted content
}
