% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twCompliance.R
\name{start_compliance_check}
\alias{start_compliance_check}
\title{Start a compliance check}
\usage{
start_compliance_check(data)
}
\arguments{
\item{data}{A dataframe generated by `combine_status()`.}
}
\value{
twCompliance ID of the compliance job. Save this to check the jobs status and download the results.
}
\description{
Start a compliance check
}
