% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{start_job}
\alias{start_job}
\title{Start compliance job}
\usage{
start_job(ids, type = "tweets", verbose = TRUE, bearer_token = get_bearer())
}
\arguments{
\item{ids}{A character vector containing the status ids of the tweets that will be checked.}

\item{type}{Currently only "tweets" is allowed here.}

\item{verbose}{If FALSE the function will not print any messages.}

\item{bearer_token}{API bearer token.}
}
\value{
string ID of the compliance job. Save this to check the jobs status and download the results.
}
\description{
This function attempts to start a compliance job with a given list of tweet status ids.
}
