% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_job.R
\name{download_job}
\alias{download_job}
\title{Download compliance job}
\usage{
download_job(job_id)
}
\arguments{
\item{job_id}{A string containing the ID of the compliance job that will be downloaded.}
}
\value{
dataframe A dataframe of all returned tweets.
}
\description{
This function attempts to download a compliance job if it is ready to download.
}
